+++ mail-rs/src/utils/mod.rs
@@ -1,4 +1,6 @@
+pub mod dkim;
 pub mod dns;
 pub mod email;
+pub mod spf;
 
 pub use email::validate_email;